Added the IntegrationTests project and wrote two simple tests for GmailApi.GetEmail(s). Slight refactoring to GlobalValues.cs and Extensions.cs
