+++ src/ADHDmailTests/ADHDmailIntegrationTests/packages.config
@@ -1,5 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="Microsoft.CodeCoverage" version="15.9.0" targetFramework="net461" />
   <package id="xunit" version="2.4.1" targetFramework="net461" />
   <package id="xunit.abstractions" version="2.0.3" targetFramework="net461" />
   <package id="xunit.analyzers" version="0.10.0" targetFramework="net461" />
@@ -7,4 +8,5 @@
   <package id="xunit.core" version="2.4.1" targetFramework="net461" />
   <package id="xunit.extensibility.core" version="2.4.1" targetFramework="net461" />
   <package id="xunit.extensibility.execution" version="2.4.1" targetFramework="net461" />
+  <package id="xunit.runner.visualstudio" version="2.4.1" targetFramework="net461" developmentDependency="true" />
 </packages>
