+++ src/ADHDmailTests/ADHDmailUnitTests/packages.config
@@ -1,12 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Microsoft.Bcl" version="1.1.10" targetFramework="net461" />
-  <package id="Microsoft.Bcl.Build" version="1.0.21" targetFramework="net461" />
-  <package id="Microsoft.Bcl.Compression" version="3.9.85" targetFramework="net461" />
   <package id="Microsoft.CodeCoverage" version="15.9.0" targetFramework="net461" />
-  <package id="Microsoft.Net.Http" version="2.2.29" targetFramework="net461" />
-  <package id="Microsoft.NET.Test.Sdk" version="15.9.0" targetFramework="net461" />
-  <package id="NUnit" version="3.11.0" targetFramework="net461" />
   <package id="xunit" version="2.4.1" targetFramework="net461" />
   <package id="xunit.abstractions" version="2.0.3" targetFramework="net461" />
   <package id="xunit.analyzers" version="0.10.0" targetFramework="net461" />
